+++ src/interactive/details/history/mod.rs
@@ -67,6 +67,7 @@ fn draw_table(
         let time = entry.time.to_string();
         let qos = format::qos(entry.qos).to_owned();
         let value = match &entry.payload {
+            Payload::Binary(data) => format!("{data:?}"),
             Payload::Json(json) => JsonSelector::get_json(json, json_selector)
                 .unwrap_or(json)
                 .to_string(),
@@ -75,7 +76,6 @@ fn draw_table(
                     .unwrap_or(messagepack)
                     .to_string()
             }
-            Payload::NotUtf8(err) => format!("invalid UTF-8: {err}"),
             Payload::String(str) => str.to_string(),
         };
         let row = Row::new(vec![time, qos, value]);